pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "streamlit-folium>=0.24.0",
    "streamlit>=1.43.2",
    "tensorflow>=2.14.0",
    "altair>=5.5.0",
    "pandas>=2.2.3",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "folium>=0.19.5",
]
